Refactor shims to avoid duplicating code
